feat: tests, cred schema validation, schema fixes
