+++ src/credentials/types/index.ts
@@ -1,0 +1,3 @@
+import { W3CCredential } from 'did-jwt-vc';
+
+export type CredentialType = W3CCredential;
